requirements.txt
pydantic
openpyxl
pillow
ultralytics
fastapi
asyncpg
uvicorn
loguru
python-dotenv
boto3